[step3] feature 보너스 볼 관련 기능 구현

+++ src/main/java/lottosecond/domain/EarningRateCalculator.java
@@ -5,7 +5,7 @@ import lottosecond.domain.lotto.Lottos;
 public class EarningRateCalculator {
 
     public double calculateEarningRate(WinnerBoard winnerBoard, Lottos lottos) {
-        int totalPrice = winnerBoard.calculateTotalPrice();
+        long totalPrice = winnerBoard.calculateTotalPrice();
         return Math.floor(totalPrice * 100.0 / lottos.totalLottoBuyMoney()) / 100;
     }
 }
